Family Details Section (Work in Progress)
The family details section of the admission form is nearly complete, with the exception of the conditional logic for the occupation combobox. Currently, if the user selects "Other" as the occupation, the expected input field does not appear. This feature is still under development, and once implemented, users will be able to provide custom input when choosing "Other" as their occupation option.